requirements.txt
snap-stanford